requirements.txt
Django==4.1.5
djangorestframework==3.14.0
gunicorn==20.1.0
whitenoise==5.3.0
django-cors-headers==3.11.0
django-filter